+++ src/main/java/lotto/domain/price/PurchaseInfo.java
@@ -1,6 +1,5 @@
-package lotto.domain.dto;
+package lotto.domain.price;
 
-import lotto.domain.price.Price;
 import lotto.domain.ticket.LottoTickets;
 import lotto.exception.AvailableCountExceedException;
 import lotto.exception.ErrorMessage;
